Exclude filestyem package from updating
With the release of centos 8.3, we are hitting a podman failure in
trying to update the filesystem package. This is because of how rootless
containers work.  Sadly, there is nothing we can do, without running our
image build as root.

For now, exclude the package until newer centos 8.3 images are published
to dockerhub.

Signed-off-by: Paul Belanger <pabelanger@redhat.com>